refactor: move mosque map screen and rename composables and files
