+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/mosque/NearbyMosquesScreen.kt
@@ -1,8 +1,9 @@
-package net.thechance.mena.faith.presentation.feature.mosque.component
+package net.thechance.mena.faith.presentation.feature.mosque
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import io.github.dellisd.spatialk.geojson.Position
 import net.thechance.mena.faith.presentation.utils.MapStyle
 import org.maplibre.compose.camera.CameraPosition
 import org.maplibre.compose.camera.rememberCameraState
@@ -10,9 +11,11 @@ import org.maplibre.compose.map.MaplibreMap
 import org.maplibre.compose.style.BaseStyle
 
 @Composable
-fun SimpleMap(
-    initialCameraPosition: CameraPosition
-) {
+fun NearbyMosquesScreen() {
+    val initialCameraPosition = CameraPosition(
+        target = Position(longitude = 14.7749, latitude = -5.4194),
+        zoom = 14.0
+    )
     val cameraState = rememberCameraState(firstPosition = initialCameraPosition)
 
     MaplibreMap(
